backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==0.28.1
bcrypt==4.0.1 